requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.3.2
cryptography==36.0.1
execnet==1.9.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
iniconfig==1.1.1
jsonpath==0.82
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.1
pytest-assume==2.4.3
pytest-base-url==1.4.2
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-selenium==2.0.1
pytest-sugar==0.9.4
pytest-timeout==2.1.0
pytest-variables==1.9.0
pytest-xdist==2.5.0
PyYAML==6.0
requests==2.27.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tenacity==6.3.1
termcolor==1.1.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.8
wsproto==1.1.0
zipp==3.7.0
